fixed problem with declaring a variable in a case: statment...doesnt work. Apparently this is known limitation of switch/case but compiler doesn't warn. Statemachines are hard to debug with only Serial.print(), and silent failures are _very_ difficult. User interface is now working.  Continuing to test EEPROM and config options.

+++ include/SoftwareConfig.h
@@ -7,10 +7,10 @@
 
 // Public functions
 sConfig_t InitDefaultConfig();             // initialze config structure in memory
-sConfig_t         GetConfig( int address); // read config from EEPROM
-void              PutConfig( int address, sConfig_t Config); // write config if needed, update CRC16
-uint16_t        CalcCRC(sConfig_t Config);
-bool     isConfigValid(sConfig_t Config); // check config.CRC16
-void       PrintConfig(sConfig_t Config); // pretty print config on serial port
+sConfig_t GetConfig(uint8_t address); // read config from EEPROM
+void PutConfig(uint8_t address, sConfig_t Config); // write config if needed, update CRC16
+uint16_t CalcCRC(sConfig_t Config);
+bool isConfigValid(sConfig_t Config); // check config.CRC16
+void PrintConfig(sConfig_t Config); // pretty print config on serial port
 
 #endif
